backend/requirements.txt
Django==3.2.7
Djangorestframework~=3.13.1
psycopg2-binary~=2.9.3
django-cors-headers == 3.11.0